isolated checking for from_chars (AppleClang is such a headache)
