+++ src/SimpleTree.cpp
@@ -24,35 +24,13 @@
 
 #include "SimpleTree.h"
 #include "errors.h"
+#include "utilities.hpp"
 
 #include <nlohmann/json.hpp>
 
-#include <charconv> // from_chars
 #include <iostream>
-#include <type_traits> // true_type for detecting from_chars
 #include <unordered_set>
-#include <utility> // declval for detecting from_chars
 #include <vector>
-
-/// C++17 compile-time test for presence of std::from_chars(const char*, const
-/// char*, T&) Replace it with C++20 concepts later (or perhaps AppleClang will
-/// implement proper from_chars by then). History: C++17 introduced
-/// std::from_chars, but STL vendors were late, then provided only integral
-/// versions...
-template <typename, typename = void>
-struct has_from_chars : std::false_type {
-}; // primary template declaration (used when specializations fail)
-
-template <typename T>
-struct has_from_chars< // template partial specialization
-    T,
-    std::void_t< // tests if the following expression computes into a type:
-        decltype(std::from_chars(std::declval<const char *&>(),
-                                 std::declval<const char *&>(),
-                                 std::declval<T &>()))>> : std::true_type {};
-
-template <typename T>
-constexpr auto has_from_chars_v = has_from_chars<T>::value;
 
 using json = nlohmann::json;
