+++ test/CMakeLists.txt
@@ -1,6 +1,10 @@
 find_package (Boost COMPONENTS unit_test_framework REQUIRED)
 include_directories (${CMAKE_SOURCE_DIR}/src)
 
+add_executable (utilities_test utilities_test.cpp)
+add_test(NAME utilities_test COMMAND utilities_test)
+
+
 add_executable (SimpleTree_test SimpleTree_test.cpp)
 target_link_libraries(SimpleTree_test PRIVATE SimpleTree Boost::unit_test_framework)
 add_test(NAME SimpleTree_test COMMAND SimpleTree_test)
